+++ node/bindnode/node.go
@@ -3,177 +3,11 @@ package bindnode
 import (
 	"fmt"
 	"reflect"
-	"strings"
 
 	ipld "github.com/ipld/go-ipld-prime"
 	basicnode "github.com/ipld/go-ipld-prime/node/basic"
 	"github.com/ipld/go-ipld-prime/schema"
 )
-
-// WrapNoSchema implements an ipld.Node given a pointer to a Go value.
-//
-// Same rules as PrototypeNoSchema apply.
-func WrapNoSchema(ptr interface{}) ipld.Node {
-	panic("TODO")
-	// ptrVal := reflect.ValueOf(ptr)
-	// if ptrVal.Kind() != reflect.Ptr {
-	// 	panic("must be a pointer")
-	// }
-	// return &_node{val: ptrVal.Elem()}
-}
-
-// Unwrap takes an ipld.Node implemented by one of the Wrap* or Prototype* APIs,
-// and returns a pointer to the inner Go value.
-//
-// Unwrap returns the input node if the node isn't implemented by this package.
-func Unwrap(node ipld.Node) (ptr interface{}) {
-	var val reflect.Value
-	switch node := node.(type) {
-	case *_node:
-		val = node.val
-	case *_nodeRepr:
-		val = node.val
-	default:
-		return node
-	}
-	if val.Kind() == reflect.Ptr {
-		panic("didn't expect val to be a pointer")
-	}
-	if !val.CanAddr() {
-		// Not addressable? Just return the interface as-is.
-		// TODO: This happens in some tests, figure out why.
-		return val.Interface()
-	}
-	return val.Addr().Interface()
-}
-
-// PrototypeNoSchema implements an ipld.NodePrototype given a Go pointer type.
-//
-// In this form, no IPLD schema is used; it is entirely inferred from the Go
-// type.
-//
-// Go types map to schema types in simple ways: Go string to schema String, Go
-// []byte to schema Bytes, Go struct to schema Map, and so on.
-//
-// A Go struct field is optional when its type is a pointer. Nullable fields are
-// not supported in this mode.
-func PrototypeNoSchema(ptrType interface{}) ipld.NodePrototype {
-	panic("TODO")
-	// typ := reflect.TypeOf(ptrType)
-	// if typ.Kind() != reflect.Ptr {
-	// 	panic("must be a pointer")
-	// }
-	// return &_prototype{goType: typ.Elem()}
-}
-
-// PrototypeOnlySchema implements an ipld.NodePrototype given an IPLD schema type.
-//
-// In this form, Go values are constructed with types inferred from the IPLD
-// schema, like a reverse of PrototypeNoSchema.
-func PrototypeOnlySchema(schemaType schema.Type) ipld.NodePrototype {
-	goType := inferGoType(schemaType)
-	return prototype(goType, schemaType)
-}
-
-// from IPLD Schema field names like "foo" to Go field names like "Foo".
-func fieldNameFromSchema(name string) string {
-	return strings.Title(name)
-}
-
-func inferGoType(typ schema.Type) reflect.Type {
-	switch typ := typ.(type) {
-	case *schema.TypeBool:
-		return goTypeBool
-	case *schema.TypeInt:
-		return goTypeInt
-	case *schema.TypeFloat:
-		return goTypeFloat
-	case *schema.TypeString:
-		return goTypeString
-	case *schema.TypeBytes:
-		return goTypeBytes
-	case *schema.TypeStruct:
-		fields := typ.Fields()
-		goFields := make([]reflect.StructField, len(fields))
-		for i, field := range fields {
-			ftyp := inferGoType(field.Type())
-			if field.IsNullable() {
-				ftyp = reflect.PtrTo(ftyp)
-			}
-			if field.IsOptional() {
-				ftyp = reflect.PtrTo(ftyp)
-			}
-			goFields[i] = reflect.StructField{
-				Name: fieldNameFromSchema(field.Name()),
-				Type: ftyp,
-			}
-		}
-		return reflect.StructOf(goFields)
-	case *schema.TypeMap:
-		ktyp := inferGoType(typ.KeyType())
-		vtyp := inferGoType(typ.ValueType())
-		if typ.ValueIsNullable() {
-			vtyp = reflect.PtrTo(vtyp)
-		}
-		// We need an extra field to keep the map ordered,
-		// since IPLD maps must have stable iteration order.
-		// We could sort when iterating, but that's expensive.
-		// Keeping the insertion order is easy and intuitive.
-		//
-		//	struct {
-		//		Keys   []K
-		//		Values map[K]V
-		//	}
-		goFields := []reflect.StructField{
-			{
-				Name: "Keys",
-				Type: reflect.SliceOf(ktyp),
-			},
-			{
-				Name: "Values",
-				Type: reflect.MapOf(ktyp, vtyp),
-			},
-		}
-		return reflect.StructOf(goFields)
-	case *schema.TypeList:
-		etyp := inferGoType(typ.ValueType())
-		if typ.ValueIsNullable() {
-			etyp = reflect.PtrTo(etyp)
-		}
-		return reflect.SliceOf(etyp)
-	case *schema.TypeUnion:
-		// We need an extra field to record what member we stored.
-		type goUnion struct {
-			Index int // 0..len(typ.Members)-1
-			Value interface{}
-		}
-		return reflect.TypeOf(goUnion{})
-	}
-	panic(fmt.Sprintf("%T\n", typ))
-}
-
-// Prototype implements an ipld.NodePrototype given a Go pointer type and an
-// IPLD schema type.
-//
-// In this form, it is assumed that the Go type and IPLD schema type are
-// compatible. TODO: check upfront and panic otherwise
-func Prototype(ptrType interface{}, schemaType schema.Type) ipld.NodePrototype {
-	goPtrType := reflect.TypeOf(ptrType)
-	if goPtrType.Kind() != reflect.Ptr {
-		panic("ptrType must be a pointer")
-	}
-	return prototype(goPtrType.Elem(), schemaType)
-}
-
-func prototype(goType reflect.Type, schemaType schema.Type) ipld.NodePrototype {
-	if goType.Kind() == reflect.Invalid {
-		panic("goType must be valid")
-	}
-	if schemaType == nil {
-		panic("schemaType must not be nil")
-	}
-	return &_prototype{schemaType: schemaType, goType: goType}
-}
 
 // Assert that we implement all the interfaces as expected.
 // Grouped by the interfaces to implement, roughly.
